29.03.2020
убрал stepper не работал, поменял изображение добавил изображение приложения, пофиксил таб бар

+++ modelForStepper.swift
@@ -5,7 +5,7 @@ class BuyStat {
     
     var purchaseAmount = 0
     
-    init(purchaseAmount : Double) {
+    init(purchaseAmount : Int) {
         self.purchaseAmount = Int(purchaseAmount)
     }
 }
